pyproject.toml
[tool.poetry]
name = "auto-conso"
version = "0.1.0"
description = "Client web pour enregistrer les pertes et auto consomations"
authors = ["Vincent Bisserie <vbisserie@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-ldap = "^3.3.1"
flask-jwt-extended = "^4.1.0"
flask-cors = "^3.0.10"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
